Restructure uplink as a library
